[Serializer] Use PropertyAccessor to access / mutate values in AbstractObjectNormalizer
